Configs cleaned up and conflicts removed
This patch and the new preliminary RU support were developed simultaneously, so this update uses the IDs that RU will use for its 1.6.* support.

IDs are mostly identical to the 1.5.2 support but not completely due to new mods and fresh conflicts.

NO ATTEMPT TO BALANCE THE MODS WAS MADE IN THIS COMMIT - THIS COMMIT IS DESIGNED SOLELY TO GET THE DEV VERSION TO LOAD AT ALL.