Further updates and patches for RR 2.0.8
This adds CustomStuff2 which allows you to add mods in Javascript. More limited but much more transparent and good for the quick patch jobs I was using it for.

Added is a custom recipe patcher to resolve the uncraftable Atomic Science plates issue. A placeholder conversion recipe between Refined Iron Plates and UE Steel Plates has also been implemented.

The rest of this patch is mod updates and recipe clash corrections that didn't require new recipes.