Worldgen corrections for Mariculture and TiCo
Bauxite has been banded at last (thanks Joshie!) and due to redundancy all TiCo aluminium has been pulled - rumour has it that it's about to become Zinc anyway (which is great news for us!)

Includes extra config cleanup. This alone could be released as a hotfix to 2.4.0, but we have other work to do today. :)

Also, after forgetting about it for ages, I remembered to restore the coloured wood in ExtraUtils - the bug we were working around should be long gone!

Finally I added the Metallurgy+Friends oregen heightmap text file to the distro, as it's ended up being so crucial to the pack.