Stealth patch to MineTweaker for server compatibility
Looks like "vat" is treated as a reserved word by MineTweaker, causing shenanigans.

Additionally, for servers the MineTweaker scripts have to be placed in the save games folder (which defaults to "world"), therefore a ZIP is now included with the server version of the pack that adds the MineTweaker configs to the right location.